requirements.txt
anthropic>=0.40.0
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
db-dtypes>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tabulate>=0.9.0
python-dotenv>=1.0.0
